Updated the quit script to check if the output file exists

+++ scripts/quit.sh
@@ -2,6 +2,7 @@
 
 cd data
 
+# Checks for old data files and removes if present
 if [ -f bookData.txt ]; then
     rm bookData.txt
 fi
@@ -9,5 +10,11 @@ if [ -f tagData.txt ]; then
     rm tagData.txt
 fi
 
-mv bookOutput.txt bookData.txt
-mv tagOutput.txt tagData.txt
+# Takes current status and saves it as the input
+# for the next run of the program
+if [ -f bookOutput.txt ]; then
+    mv bookOutput.txt bookData.txt
+fi
+if [ -f tagOutput.txt ]; then
+    mv tagOutput.txt tagData.txt
+fi
